Still updating to 0.9.7. We're getting close to being completely updated! Only 20 or so more files. They also just happen to be the most significant. Oh boy.
Some rather major changes in this one. Not major as in "you'll notice this" but rather major as in "Vanilla changed some potentially dangerous variables but we actually got them to work with our stuff".

+++ trunk/source/src/NDSSystem.cpp
@@ -261,7 +261,7 @@ NDS_header * NDS_getROMHeader(void)
 	memcpy(header->logo, MMU.CART_ROM + 192, 156);
 	header->logoCRC16 = T1ReadWord(MMU.CART_ROM, 348);
 	header->headerCRC16 = T1ReadWord(MMU.CART_ROM, 350);
-	memcpy(header->reserved, MMU.CART_ROM + 352, min(160, gameInfo.romsize - 352));
+	memcpy(header->reserved, MMU.CART_ROM + 352, min(160, (int)gameInfo.romsize - 352));
 
 	return header;
 }
